displaying data

+++ FindRepo/Core/Entities/Owner.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Owner: Codable, Identifiable {
-    let id: Int
+    let id: Int64
     let name: String
     let avatarUrl: String
     let profileUrl: String
